Isolate state enumeration tests
