+++ openff/toolkit/_tests/molecule/__init__.py
@@ -1,0 +1,1 @@
+"""Tests of the Molecule API."""
